requirements.txt
requests
jupyter
pytest==7.1.3
pytest-pycodestyle
flask
pandas
cerberus
requests
numpy
Pillow
matplotlib
pymodm
dnspython


